Timer in its own bloc
